feat: Add dedicated dialogs for updating user profile and workspace settings, refactoring their respective tabs.
